requirements.txt
aiogram>=3.7.0,<4.0.0
httpx>=0.27.0,<1.0.0
qrcode>=7.4.2,<8.0.0
apscheduler>=3.10.4,<4.0.0
python-dotenv>=1.0.1,<2.0.0
